Add own file for metadata; Add GetMetaData func for retrieving updateset metadata.
